Alternative to Prestissimo [4.1+] and other annoyances

+++ core/src/main/java/de/danoeh/antennapod/core/storage/PodDBAdapter.java
@@ -309,7 +309,11 @@ public class PodDBAdapter {
             dbHelper.close();
             dbHelper = null;
         }
-        return context.deleteDatabase(PodDBAdapter.DATABASE_NAME);
+        if(context != null) { // may not have been initialized
+            return context.deleteDatabase(PodDBAdapter.DATABASE_NAME);
+        } else {
+            return false;
+        }
     }
 
     /**
